tidy up

+++ app/src/main/java/com/pixels/blockies/app/draws/BlockDrawable.java
@@ -5,16 +5,50 @@ import android.graphics.Paint;
 import com.pixels.blockies.app.draws.api.Drawable;
 import com.pixels.blockies.app.draws.enums.GameColor;
 
+/**
+ * Draws a single block on a specific location by given width and height.
+ * It can also be adjusted (shrink/expand) by a factor. That way the proportions
+ * to its stroke thickness will be kept.
+ */
 class BlockDrawable implements Drawable {
+
     private Paint paint = new Paint();
+
+    /**
+     * The cooridinates of the block
+     */
     private int x = 0;
     private int y = 0;
+
+    /**
+     * The color of the block
+     */
     private int specificColor = -1;
+
+    /**
+     * The stroke size of the block
+     */
     private int specificBlockStroke = -1;
-    private int adjustment = -1;
+
+    /**
+     * The adjustment to scale the block up or down
+     */
+    private float adjustment = -1;
+
+    /**
+     * Width and height of the block
+     */
     private int blockWidth = -1;
     private int blockHeight = -1;
 
+    /**
+     * Blockdrawable that keeps its size as defined by blockwidth and blockheight
+     *
+     * @param x
+     * @param y
+     * @param blockwidth
+     * @param blockHeight
+     */
     public BlockDrawable(int x, int y, int blockwidth, int blockHeight){
         this.x = x;
         this.y = y;
@@ -23,7 +57,16 @@ class BlockDrawable implements Drawable {
         this.blockHeight = blockHeight;
     }
 
-    public BlockDrawable(int x, int y, int blockwidth, int blockHeight, int adjustment){
+    /**
+     * Blockdrawable that adjusts its size and stroke by a given factor
+     *
+     * @param x
+     * @param y
+     * @param blockwidth
+     * @param blockHeight
+     * @param adjustment
+     */
+    public BlockDrawable(int x, int y, int blockwidth, int blockHeight, float adjustment){
         this.x = x;
         this.y = y;
         this.adjustment = adjustment;
